Add settings INLINE_MEDIA_REMOVE_TAGS to remove unexpected html tags.
As a result of using 'html.parser' as the parser for BeautifulSoup4
news tags are added to the result of parsing TextFieldWithInlines.
There are other parsers but none available for Python 3.

The new setting defines a list of tags to be remove. It's mean to
list the complete tag, not a regular expression, so only matching
tags will be removed. It defaults to ['</br>']